+++ ThorSoft.Optics/ThorSoft.Optics.Generator/FocusProperties/FocusPropertiesLens.cs
@@ -1,0 +1,9 @@
+﻿namespace ThorSoft.Optics.Generator.FocusProperties
+{
+    internal sealed record class Lens
+    {
+        public required string Name { get; init; }
+        public required string Type { get; init; }
+        public required string Visibility { get; init; }
+    }
+}
