+++ ThorSoft.Optics/ThorSoft.Optics.Generator/FocusProperties/FocusPropertiesOutput.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.CodeAnalysis;
+using ThorSoft.Optics.Generator.Util;
+
+namespace ThorSoft.Optics.Generator.FocusProperties
+{
+    internal sealed record class FocusPropertiesOutput
+    {
+        public string? TypeName { get; init; }
+        public string? TypeKind { get; init; }
+
+        public EquatableMemory<Diagnostic> Diagnostics { get; init; }
+        public EquatableMemory<Lens> Lenses { get; init; }
+    }
+}
