+++ ThorSoft.Optics/ThorSoft.Optics.Generator/OpticsGenerator.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.CodeAnalysis;
 using ThorSoft.Optics.Generator.FocusProperties;
-using ThorSoft.Optics.Generator.Util;
 
 namespace ThorSoft.Optics.Generator
 {
@@ -28,30 +27,5 @@ namespace ThorSoft.Optics.Generator
                 focusProperties,
                 FocusPropertiesGeneration.Invoke);
         }
-    }
-
-    internal sealed record class FocusPropertiesOutput
-    {
-        public string? TypeName { get; init; }
-        public string? TypeKind { get; init; }
-
-        public EquatableMemory<Diagnostic> Diagnostics { get; init; }
-        public EquatableMemory<Lens> Lenses { get; init; }
-    }
-
-    internal sealed record class Lens
-    {
-        public required string Name { get; init; }
-        public required string Type { get; init; }
-        public required string Visibility { get; init; }
-    }
-
-    file static class Extensions
-    {
-        public static FocusPropertiesOutput AsOutput(this Diagnostic diagnostic) =>
-            new()
-            {
-                Diagnostics = new EquatableMemory<Diagnostic>([diagnostic])
-            };
     }
 }
